update
Default constructor:
Some default values are initialized to prevent null pointer exceptions.
Parameterized constructors:
Use the setter method to set instance variables to ensure the validity of the data.
Getter & Setter:
Input validation has been added to ensure that the incoming string is not empty.
Use the trim() method to remove whitespace characters at the beginning and end of the string.
How to print details:
Added a check to see if doctor is null to avoid null pointer exceptions.
toString method:
A string representation of the object is provided for easy debugging and printing.